requirements.txt
grpcio
torch
numpy
marisa_trie
